js文件名加testlink用例编号 修复分区表用例没执行的bug,增加所有读lob的结果校验.
git-svn-id: http://192.168.20.11/sequoiadb/trunk@33402 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ testcase_new/story/js/lob/lob_laterHashSplit_normalRecord_13942.js
@@ -7,14 +7,19 @@
 
 function main( db )
 {
-   var testFile = CHANGEDPREFIX + "lobTest.file",
-       getTestFile = CHANGEDPREFIX + "lobTestGet.file",
-       putNum = 100,
-       oid = new Array(),
-       partitionNum = 2048,
-       cmd = new Cmd() ;
-
-   lobAutoFile( testFile ) ;   // auto file
+   var testFile = CHANGEDPREFIX + "lobTest.file" ;
+   var getTestFile = CHANGEDPREFIX + "lobTestGet.file" ;
+   var putNum = 50;
+   var partitionNum = 2048 ;
+   
+   var names = lobGetAllGroupNames( db ) ;
+   if( 1 == names.length )
+   {
+      return ;
+   }
+   
+   lobGenerateFile( testFile ) ;   // auto file
+   var originMd5 = getMd5ForFile( testFile ) ;
    // create collection
    var optionObj = { "ShardingKey":{"no":1}, "ShardingType":"hash", "ReplSize":0,
                      "Partition":partitionNum, "Compressed":true } ;
@@ -23,8 +28,8 @@ function main( db )
    // put normal data and lob data
    try
    {
-      lobInsert( cl, putNum ) ;
-      oid = lobPutLob( cl, testFile, putNum ) ;
+      lobInsertDoc( cl, putNum ) ;
+      var oids = lobPutLob( cl, testFile, putNum ) ;
       println( "success to put normal data and lob data" ) ;
    }
    catch( e )
@@ -36,48 +41,31 @@ function main( db )
    // do hash split collection here
    try
    {
-      if( false != commIsStandalone(db) )
-         throw "run mode is standalone" ;
       var FULLCLNAME = COMMCSNAME + "." + COMMCLNAME ;
-      var clRg = commGetCLGroups( db, FULLCLNAME ),
-          rg = lobGetGroups( db ) ;
-      if( 1 != clRg.length )
-      {
-         println( "collection have more than one groups: " + clRg ) ;
-         throw "NotOneGroup" ;
-      }
-      if( 1 != rg.length )
-         throw "database only have one group" ;
-      var cond = partitionNum/rg.length ;
+      var clRg = commGetCLGroups( db, FULLCLNAME ) ;
+      var cond = Math.floor(partitionNum/names.length) ;
       //pritnln( "the group length: " + cond ) ;
       var loopCond = cond ;
-      for( var i = 0 ; i < rg.length ; ++i )
+      for( var i = 0 ; i < names.length ; ++i )
       {
-         if( clRg[0] != rg[i] )
+         if( clRg[0] != names[i] )
          {
             var firstCond = { "Partition": (loopCond-cond) } ;
             var secondCond = { "Partition": loopCond } ;
-            lobSplit( cl, clRg[0], rg[i], firstCond, secondCond ) ;
+            lobSplit( cl, clRg[0], names[i], firstCond, secondCond ) ;
             loopCond += cond ;
          }
       }
       println( "success to split collection" ) ;
-   }
-   catch( e )
-   {
-      if( "run mode is standalone" != e &&
-          "database only have one group" != e )
+      
+      for( var i = 0 ; i < oids.length ; ++i )
       {
-         println( "failed to split collection, rc = " + e ) ;
-         throw e ;
-      }
-   }
-   // get lob
-   try
-   {
-      for( var i = 0 ; i < oid.length ; ++i )
-      {
-         cl.getLob( oid[i], getTestFile, true ) ;
+         cl.getLob( oids[i], getTestFile, true ) ;
+         var curMd5 = getMd5ForFile( getTestFile ) ;
+         if ( originMd5 !== curMd5 )
+         {
+            throw "origin file's md5=" + originMd5 + "getLob's md5=" + curMd5 ;
+         }
       }
       println( "success to get lob" ) ;
       for( var i = 0 ; i < cl.count() ; ++i )
@@ -90,34 +78,41 @@ function main( db )
          }
       }
       println( "success to query" ) ;
-      // remove file
-      cmd.run( "rm -rf " + testFile ) ;
-      cmd.run( "rm -rf " + getTestFile ) ;
    }
    catch( e )
    {
       // remove file
-      cmd.run( "rm -rf " + testFile ) ;
-      cmd.run( "rm -rf " + getTestFile ) ;
       println( "failed to get lob and query nomral data, rc = " + e ) ;
       throw e ;
+   }
+   finally
+   {
+      var cmd = new Cmd() ;
+      cmd.run( "rm -rf " + testFile ) ;
+      if ( lobFileIsExist( getTestFile ) )
+      {
+         cmd.run( "rm -rf " + getTestFile ) ;
+      }
    }
 }
 
 // Run Main
 try
 {
-   commDropCL( db, COMMCSNAME, COMMCLNAME, true, true,
-               "clear collection in the beginning" ) ;
-   main( db ) ;
-   commDropCL( db, COMMCSNAME, COMMCLNAME, false, false,
-               "drop collection in the end, correct" ) ;
-   db.close( ) ;
+   if ( !commIsStandalone(db) )
+   {
+      commDropCL( db, COMMCSNAME, COMMCLNAME, true, true,
+                  "clear collection in the beginning" ) ;
+      main( db ) ;
+   }
 }
 catch( e )
 {
-   commDropCL( db, COMMCSNAME, COMMCLNAME, true, true,
-               "drop collection in the end , error" ) ;
-   db.close( ) ;
    throw e ;
+}
+finally
+{
+   commDropCL( db, COMMCSNAME, COMMCLNAME, true, true,
+               "drop collection in the end, correct" ) ;
+   db.close( ) ;
 }
